Moved event handler for saving image/data
